Add files via upload
Revantusk Village Grind Profile for Horde 1.0.0

So for this profile You start off in Revantusk in the Hinterlands in the Eastern Kindoms.

I used this profile to get from level 43 to 50 myself and tested it personally. This profile was tested with WRobot 2.8.0 (26869) only with a Horde character.

The mobs here range from about 44 to 51 there is a mini boss at the back of this profile which is a turtle that is level 51 at the back of this profile.

It is also good for harvesting Thick levels as well off of the Saltwater Snapjaws here as well.

It is also good for harvesting Zesty Clam Meat as well from the big mouth clams as well. you will also randomly get Golden pearls Iridescent Pearls and Black Pearls randomly from the big mouth clams as well.

For the Revantusk Village The Hinterlands 43 - 50 I would recommend the plugin Free] Air Breather(Swimming) 1.0.0 because this profile will take you to into the water and there is a chance you may drown without it.

This is my first attempt at creating a profile so be nice to me.

The Free] Air Breather(Swimming) 1.0.0 can be found at https://wrobot.eu/files/file/1925-free-air-breatherswimming/

For any technical support regarding this plugin pleaase contact the author as I am not the author.

This was only tested with a Horde character on a 3.3.5a server that is "Blizlike" (Wrath of the Lich King).